chore: re-organize the middleware
