+++ Makefile
@@ -16,7 +16,7 @@ docker:
 	docker build -t go-clean-arch .
 
 run:
-	docker-compose up -d
+	docker-compose up --build -d
 
 stop:
 	docker-compose down
@@ -26,12 +26,6 @@ lint-prepare:
 	curl -sfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh| sh -s latest
 
 lint:
-	./bin/golangci-lint run \
-		--exclude-use-default=false \
-		--enable=golint \
-		--enable=gocyclo \
-		--enable=goconst \
-		--enable=unconvert \
-		./...
+	./bin/golangci-lint run ./...
 
 .PHONY: clean install unittest build docker run stop vendor lint-prepare lint
