+++ app/main.go
@@ -5,18 +5,17 @@ import (
 	"fmt"
 	"log"
 	"net/url"
-	"os"
 	"time"
 
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/labstack/echo"
 	"github.com/spf13/viper"
 
-	_articleHttpDeliver "github.com/bxcodec/go-clean-arch/article/delivery/http"
+	_articleHttpDelivery "github.com/bxcodec/go-clean-arch/article/delivery/http"
+	_articleHttpDeliveryMiddleware "github.com/bxcodec/go-clean-arch/article/delivery/http/middleware"
 	_articleRepo "github.com/bxcodec/go-clean-arch/article/repository/mysql"
 	_articleUcase "github.com/bxcodec/go-clean-arch/article/usecase"
 	_authorRepo "github.com/bxcodec/go-clean-arch/author/repository/mysql"
-	"github.com/bxcodec/go-clean-arch/middleware"
 )
 
 func init() {
@@ -27,7 +26,7 @@ func init() {
 	}
 
 	if viper.GetBool(`debug`) {
-		fmt.Println("Service RUN on DEBUG mode")
+		log.Println("Service RUN on DEBUG mode")
 	}
 }
 
@@ -43,8 +42,9 @@ func main() {
 	val.Add("loc", "Asia/Jakarta")
 	dsn := fmt.Sprintf("%s?%s", connection, val.Encode())
 	dbConn, err := sql.Open(`mysql`, dsn)
-	if err != nil && viper.GetBool("debug") {
-		fmt.Println(err)
+
+	if err != nil {
+		log.Fatal(err)
 	}
 	err = dbConn.Ping()
 	if err != nil {
@@ -59,14 +59,14 @@ func main() {
 	}()
 
 	e := echo.New()
-	middL := middleware.InitMiddleware()
+	middL := _articleHttpDeliveryMiddleware.InitMiddleware()
 	e.Use(middL.CORS)
 	authorRepo := _authorRepo.NewMysqlAuthorRepository(dbConn)
 	ar := _articleRepo.NewMysqlArticleRepository(dbConn)
 
 	timeoutContext := time.Duration(viper.GetInt("context.timeout")) * time.Second
 	au := _articleUcase.NewArticleUsecase(ar, authorRepo, timeoutContext)
-	_articleHttpDeliver.NewArticleHandler(e, au)
+	_articleHttpDelivery.NewArticleHandler(e, au)
 
 	log.Fatal(e.Start(viper.GetString("server.address")))
 }
